requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
Markdown==3.7
openai==1.62.0
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tk==0.1.0
tkinterdnd2==0.4.2
tkinterweb==3.25.17
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
